+++ elpi_REPL.ml
@@ -76,8 +76,9 @@ let usage =
 ;;
 
 (* For testing purposes we declare an identity quotation *)
+let quotations = API.Quotation.new_quotations_descriptor ()
 let _ =
-  API.Quotation.register_named_quotation ~name:"elpi"
+  API.Quotation.register_named_quotation ~descriptor:quotations ~name:"elpi"
     API.Quotation.lp
 
 let _ =
@@ -138,6 +139,7 @@ let _ =
   let elpi =
     API.Setup.init
       ~legacy_parser:!legacy_parser
+      ~quotations
       ~flags:(API.Compile.to_setup_flags flags)
       ~builtins:[Builtin.std_builtins]
       ~file_resolver:(API.Parse.std_resolver ~paths ())