+++ src/compiler.mli
@@ -20,7 +20,7 @@ exception CompileError of Loc.t option * string
 type builtins = string * Data.BuiltInPredicate.declaration list
 
 type header
-val header_of_ast : flags:flags -> parser:(module Parse.Parser) -> builtins list -> Ast.Program.t ->  header
+val header_of_ast : flags:flags -> parser:(module Parse.Parser) -> State.descriptor -> QuotationHooks.descriptor -> HoasHooks.descriptor -> builtins list -> Ast.Program.t ->  header
 
 type program
 val program_of_ast : flags:flags -> header:header -> Ast.Program.t -> program
@@ -46,11 +46,7 @@ val pp_goal : (pp_ctx:pp_ctx -> depth:int -> Format.formatter -> term -> unit) -
 
 val lookup_query_predicate : program -> string -> program * Data.constant
 
-type quotation = depth:int -> State.t -> Loc.t -> string -> State.t * term
-val set_default_quotation : quotation -> unit
-val register_named_quotation : name:string -> quotation -> unit
-
-val lp : quotation
+val lp : QuotationHooks.quotation
 
 val is_Arg : State.t -> term -> bool
 val get_Args : State.t -> term StrMap.t
@@ -69,9 +65,6 @@ val static_check :
   'a query -> bool
 
 module CustomFunctorCompilation : sig
-
-  val declare_singlequote_compilation : string -> (State.t -> F.t -> State.t * term) -> unit
-  val declare_backtick_compilation : string -> (State.t -> F.t -> State.t * term) -> unit
 
   val compile_singlequote : State.t -> F.t -> State.t * term
   val compile_backtick : State.t -> F.t -> State.t * term